requirements.txt
gradio
pandas
numpy
python-dotenv
qdrant-client
langchain-qdrant
langchain-huggingface
sentence-transformers
